Long -> Integer in models' ids

+++ src/main/java/com/musikrec/musikrec/Authentication/AuthenticationController.java
@@ -33,7 +33,7 @@ public class AuthenticationController {
     }
 
     @GetMapping("/user-id")
-    public ResponseEntity<Long> getUserIdFromLogin(){
+    public ResponseEntity<Integer> getUserIdFromLogin(){
         return new ResponseEntity<>(userService.getUserIdFromLogin(), HttpStatus.ACCEPTED);
     }
 }
