+++ src/main/java/com/musikrec/musikrec/Repositories/PlaylistRepository.java
@@ -11,16 +11,16 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PlaylistRepository extends JpaRepository<Playlist, Long> {
+public interface PlaylistRepository extends JpaRepository<Playlist, Integer> {
 
     @Query("select new com.musikrec.musikrec.Dto.Responses.PlaylistResponseDto(" +
             "p.id, " +
             "p.name, " +
             "p.description) " +
             "from Playlist p where p.appUser.id = :userId")
-    List<PlaylistResponseDto> getAllPlaylistsForUser(@Param("userId") Long userId);
+    List<PlaylistResponseDto> getAllPlaylistsForUser(@Param("userId") Integer userId);
 
     @Query("select p from Playlist p join fetch p.songs " +
             "where p.id = :playlistId and p.appUser.id = :userId")
-    PlaylistDetailsResponseDto getPlaylist(@Param("userId") Long userId, @Param("playlistId") Long playlistId);
+    PlaylistDetailsResponseDto getPlaylist(@Param("userId") Long userId, @Param("playlistId") Integer playlistId);
 }
