+++ src/main/java/com/musikrec/musikrec/Repositories/SongRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SongRepository extends JpaRepository<Song, Long> {
+public interface SongRepository extends JpaRepository<Song, Integer> {
 
     @Query("select " +
             "s.id as songId, " +
